example handler error other pre request
